financial-files-service/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
pdfplumber
pandas
openpyxl
google-generativeai
google-genai